+++ apps/src/codebridge/FileBrowser/hooks/useFileUploader.tsx
@@ -1,18 +1,30 @@
 import {sendCodebridgeAnalyticsEvent} from '@codebridge/utils/analyticsReporterHelper';
 import {useCallback} from 'react';
 
+import codebridgeI18n from '@cdo/apps/codebridge/locale';
+import {START_SOURCES} from '@cdo/apps/lab2/constants';
 import {
   useFileUploader as useLab2FileUploader,
   analyticsEvents,
 } from '@cdo/apps/lab2/hooks';
+import {getAppOptionsEditBlocks} from '@cdo/apps/lab2/projects/utils';
 import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
 import {useAppSelector} from '@cdo/apps/util/reduxHooks';
+
+import {useCodebridgeContext} from '../../codebridgeContext';
+import {DEFAULT_FOLDER_ID} from '../../constants';
+import {isDuplicateFileName} from '../../utils';
 
 export const useFileUploader: Exclude<
   typeof useLab2FileUploader,
   'sendAnalyticsEvent'
 > = args => {
   const appName = useAppSelector(state => state.lab.levelProperties?.appName);
+  const {source} = useCodebridgeContext();
+  const isStartMode = getAppOptionsEditBlocks() === START_SOURCES;
+  const validationFile = useAppSelector(
+    state => state.lab.levelProperties?.validationFile
+  );
 
   const sendAnalyticsEvent = useCallback(
     (eventName: string, payload: Record<string, string>) => {
@@ -37,5 +49,27 @@ export const useFileUploader: Exclude<
     [appName]
   );
 
-  return useLab2FileUploader({sendAnalyticsEvent, ...args});
+  const validateFileName = useCallback(
+    (fileName: string) => {
+      if (
+        isDuplicateFileName({
+          fileName,
+          folderId: DEFAULT_FOLDER_ID,
+          projectFiles: source.files,
+          isStartMode,
+          validationFile,
+        })
+      ) {
+        return codebridgeI18n.duplicateFileError({fileName});
+      }
+      return undefined;
+    },
+    [source.files, isStartMode, validationFile]
+  );
+
+  return useLab2FileUploader({
+    sendAnalyticsEvent,
+    validateFileName,
+    ...args,
+  });
 };
