+++ apps/src/lab2/hooks/useFileUploader.tsx
@@ -1,11 +1,6 @@
 import React, {useCallback, useMemo, useRef} from 'react';
 
-import {DEFAULT_FOLDER_ID} from '@cdo/apps/codebridge/constants';
 import codebridgeI18n from '@cdo/apps/codebridge/locale';
-import {isDuplicateFileName} from '@cdo/apps/codebridge/utils';
-import {START_SOURCES} from '@cdo/apps/lab2/constants';
-import {getAppOptionsEditBlocks} from '@cdo/apps/lab2/projects/utils';
-import {MultiFileSource, ProjectFile} from '@cdo/apps/lab2/types';
 
 export const enum analyticsEvents {
   UPLOAD_FAILED = 'UPLOAD_FAILED',
@@ -21,8 +16,7 @@ type FileUploaderProps = {
     callbackArgs?: unknown
   ) => void;
   errorCallback: (error: string, callbackArgs?: unknown) => void;
-  source: MultiFileSource;
-  validationFile: ProjectFile | undefined;
+  validateFileName?: (fileName: string) => string | undefined;
   multiple?: boolean;
   validMimeTypes?: string[];
   sendAnalyticsEvent?: (
@@ -92,14 +86,12 @@ export const useFileUploader = ({
   callback,
   errorCallback,
   validMimeTypes,
-  source,
-  validationFile,
+  validateFileName = () => undefined,
   sendAnalyticsEvent = () => {},
   multiple = true,
 }: FileUploaderProps) => {
   const inputRef = useRef<HTMLInputElement>(null);
   const callbackArgs = useRef<unknown>();
-  const isStartMode = getAppOptionsEditBlocks() === START_SOURCES;
 
   const changeHandler = useCallback(() => {
     Array.from(inputRef.current?.files || []).forEach(file => {
@@ -114,19 +106,9 @@ export const useFileUploader = ({
         );
         return;
       }
-      if (
-        isDuplicateFileName({
-          fileName: file.name,
-          folderId: DEFAULT_FOLDER_ID,
-          projectFiles: source.files,
-          isStartMode,
-          validationFile,
-        })
-      ) {
-        errorCallback(
-          codebridgeI18n.duplicateFileError({fileName: file.name}),
-          callbackArgs.current
-        );
+      const fileNameErrorMessage = validateFileName(file.name);
+      if (fileNameErrorMessage) {
+        errorCallback(fileNameErrorMessage, callbackArgs.current);
         return;
       }
 
@@ -163,9 +145,7 @@ export const useFileUploader = ({
     });
   }, [
     validMimeTypes,
-    source.files,
-    isStartMode,
-    validationFile,
+    validateFileName,
     sendAnalyticsEvent,
     errorCallback,
     callback,
